% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultiGauss.R
\name{createCovs}
\alias{createCovs}
\title{Simulate the array of covariance parameters for the mutlivariate Gaussian mixture model. This function is adapted from https://stat.ethz.ch/pipermail/r-help/2008-February/153708}
\usage{
createCovs(k, d)
}
\arguments{
\item{k}{Number of different Gaussians}

\item{d}{Number of dimensions for multivariate Gaussian mixture model}
}
\value{
mus  : A (d x k) matrix of mean parameters
}
\description{
Simulate the array of covariance parameters for the mutlivariate Gaussian mixture model. This function is adapted from https://stat.ethz.ch/pipermail/r-help/2008-February/153708
}
