% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultiGauss.R
\name{createMixProp}
\alias{createMixProp}
\title{Simulate the mixing proportions for the Gaussian mixture model}
\usage{
createMixProp(k)
}
\arguments{
\item{k}{Number of different Gaussians}
}
\value{
mix_prop : k-dimensional Vector of mixing proportions (to be normalized)
}
\description{
Simulate the mixing proportions for the Gaussian mixture model
}
