% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultiGauss.R
\name{createMus}
\alias{createMus}
\title{Simulate the centers for each Gaussian in the mutlivariate Gaussian mixture model}
\usage{
createMus(k, d)
}
\arguments{
\item{k}{Number of different Gaussians}

\item{d}{Number of dimensions for multivariate Gaussian mixture model}
}
\value{
mus  : A (d x k) matrix of mean parameters
}
\description{
Simulate the centers for each Gaussian in the mutlivariate Gaussian mixture model
}
