% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnrm_arma}
\alias{dmvnrm_arma}
\title{Function to evaluate likelihood or log-likelihood of multivariate normal using armadillo}
\usage{
dmvnrm_arma(x, mean, sigma, logd = FALSE)
}
\arguments{
\item{x}{(n x d) matrix of observed data}

\item{mean}{Row vector (d) of means}

\item{sigma}{(d x d) matrix of covariance}

\item{logd}{boolean indicating log-likelihood (true) or likelihood (false) is returned}
}
\value{
a size n vector of log-likelihoods
}
\description{
Function to evaluate likelihood or log-likelihood of multivariate normal using armadillo
}
