% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em_alg_GMM}
\alias{em_alg_GMM}
\title{Implements the EM algorithm on mvGMM data}
\usage{
em_alg_GMM(x, k, lambda, max_it, tol)
}
\arguments{
\item{x}{A (n x d) matrix of observed data}

\item{k}{The number of mixing components}

\item{lambda}{Final value for regularization parameter - will go lambda^10 to lambda. Lambda will automatically equal infinity, if not specified. This is equivalent to a standard EM algorithm}

\item{max_it}{Maximum number of iterations}

\item{tol}{Relative tolerance of likelihood at convergence}
}
\value{
A list containing the following values:
* llk     : Likelihood of final iteration
* mus     : (d x k) of estimated means
* mix_prop : size k vector of estimated mixing proportions
* covs    : (d x d x k) vectors of estimated standard deviations
* probs : Matrix (n x k) of the probabilties of observed values in each cluster
* classification Size n vector of most likely cluster for each observed value
* iter    : number of iterations
}
\description{
Implements the EM algorithm on mvGMM data
}
