% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emControl_fn}
\alias{emControl_fn}
\title{This is an Rcpp implementation of the emControl function from the mclust package. This will be used while implementing the Mclust function.}
\usage{
emControl_fn(itmax)
}
\arguments{
\item{itmax}{Maximum number of iterations for Mclust()}
}
\value{
A named list in which the names are the names of the arguments and the values are the values supplied to the arguments.
}
\description{
This is an Rcpp implementation of the emControl function from the mclust package. This will be used while implementing the Mclust function.
}
