% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Function to calculate the Mahalanobis distance. Credit: Rcpp gallery.}
\usage{
Mahalanobis(x, center, cov)
}
\arguments{
\item{x}{(n x d) matrix of observed data}

\item{center}{Row vector (d) of means}

\item{cov}{(d x d) matrix of covariance}
}
\description{
Function to calculate the Mahalanobis distance. Credit: Rcpp gallery.
}
