% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mclust_fn}
\alias{Mclust_fn}
\title{Implements the Mclust() function from library mclust.}
\usage{
Mclust_fn(x, k)
}
\arguments{
\item{x}{A (n x d) matrix of observed data}

\item{k}{The number of mixing components}
}
\value{
A list of parameter estimates that will be used for the initial parameter estimates in the EM algorithm.
}
\description{
Implements the Mclust() function from library mclust.
}
