% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.mv.gmm.R
\name{mv.dnorm}
\alias{mv.dnorm}
\title{Function to evaluate likelihoods under multivariate normal  // HL}
\usage{
mv.dnorm(x, mus, covs)
}
\arguments{
\item{x}{(n x d) matrix of observed data    // HL}

\item{mus}{d-dimensional vector of means      // HL}

\item{covs}{(d x d) matrix of covariance       // HL}
}
\value{
a size n vector of log-likelihoods     // HL
}
\description{
Function to evaluate likelihoods under multivariate normal  // HL
}
