\name{nwREM-package}
\alias{nwREM-package}
\alias{nwREM}
\docType{package}
\title{
\packageTitle{nwREM}
}
\description{
\packageDescription{nwREM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nwREM}
\packageIndices{nwREM}
 }
\author{
\packageAuthor{nwREM}

Maintainer: \packageMaintainer{nwREM}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
