% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{plot_kmeans}
\alias{plot_kmeans}
\title{Function to plot the cluster assignment from the k-means algorithm.}
\usage{
plot_kmeans(sim_kmeans, sampleMat)
}
\arguments{
\item{sim_kmeans}{Output from the k-means function}

\item{sampleMat}{An (n x d+1) matrix with the observed data and true cluster assignments}
}
\value{
A plot containing the observed data, colored by their clusters according to k-means. Estimated mean of clusters also plotted.
}
\description{
Function to plot the cluster assignment from the k-means algorithm.
}
