% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fns.R
\name{plot_robust_EM}
\alias{plot_robust_EM}
\title{Function to plot the cluster assignment from the robust EM algorithm}
\usage{
plot_robust_EM(sim_EMfit, sampleMat)
}
\arguments{
\item{sim_EMfit}{Output from the EM algorithm}

\item{sampleMat}{An (n x d+1) matrix with the observed data and true cluster assignments}
}
\value{
A plot containing the observed data, colored by their clusters according to the EM algorithm. Estimated covariance and mean of clusters also plotted.
}
\description{
Function to plot the cluster assignment from the robust EM algorithm
}
