% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMultiGauss.R
\name{simMultiGauss}
\alias{simMultiGauss}
\title{Simulate Random Samples from Multivariate Gaussian Mixture Model}
\usage{
simMultiGauss(n, mix_prop, mus, covs, perc_out)
}
\arguments{
\item{n}{Number of random samples to simulate}

\item{mix_prop}{k-dimensional Vector of mixing proportions (to be normalized)}

\item{mus}{(d x k) Matrix of mean parameters}

\item{covs}{(d x d x k) Multi-way array of covariance parameters}

\item{perc_out}{Percent (\%) of random samples that will be outliers}
}
\value{
A list containing the following values
* x : A matrix of simulated random samples from multi-dimensional Gaussian mixtures
* z : A vector of hidden variables indicating the source components
}
\description{
Simulate Random Samples from Multivariate Gaussian Mixture Model
}
